Add support for beta regular testing

+++ src/chart.js
@@ -15,12 +15,6 @@ async function dlCharts(deviceInfo) {
   let isoDate = new Date(date.getTime() - (date.getTimezoneOffset() * 60000));
   let prefixDate = isoDate.toISOString().substring(0,10).replace(/-/g, '');
 
-  let selectors = {
-    'Speedometer2': {name: "Speedometer 2.0", id: '#Speedometer_2_0_Windows_Chrome_Canary'},
-    'WebXPRT3': {name: 'WebXPRT 3', id: '#WebXPRT_3_Windows_Chrome_Canary'},
-    'Unity3D': {name: 'Unity3D2018', id: '#Unity3D2018_Windows_Chrome_Canary'},
-    'JetStream2': {name: 'JetStream2', id: '#JetStream2_Windows_Chrome_Canary'}
-  };
   let chartsDir = path.join(process.cwd(), 'charts');
   if (!fs.existsSync(chartsDir)) {
     fs.mkdirSync(chartsDir);
@@ -44,19 +38,28 @@ async function dlCharts(deviceInfo) {
   let os = "Windows";
   if (!deviceInfo['OS'].includes('Windows'))
     os = "Ubuntu";
-  const channel = "Canary";
   const browserName = "Chrome";
+
+  const browserChannel = deviceInfo['Browser'].split('-')[1];
+  const idSuffix = `_${os}_${browserName}_${browserChannel}`;
+  let selectors = {
+    'Speedometer2': {name: "Speedometer 2.0", id: '#Speedometer_2_0'},
+    'WebXPRT3': {name: 'WebXPRT 3', id: '#WebXPRT_3' },
+    'Unity3D': {name: 'Unity3D2018', id: '#Unity3D2018' },
+    'JetStream2': {name: 'JetStream2', id: '#JetStream2' }
+  };
+
   let i = 1;
   for (let workload of settings.workloads) {
     const workloadName = workload.name;
-    let chart_page_param = `?workload=${selectors[workloadName]['name']}&channel=${channel}&os=${os}&browser=${browserName}&cpu=${cpu1}&cpu=${cpu2}`;
+    let chart_page_param = `?workload=${selectors[workloadName]['name']}&channel=${browserChannel}&os=${os}&browser=${browserName}&cpu=${cpu1}&cpu=${cpu2}`;
     let chart_page_url = chart_page_host + chart_page_param;
     console.log("chart page url: ", chart_page_url);
     await page.goto(chart_page_url, { waitUntil: 'load', timeout: 60000 });
     // Leaves some time to wait for chart fully rendered
     await page.waitForTimeout(10*1000);
-    await page.waitForSelector(selectors[workloadName]['id'], {timeout: 10*1000});
-    let element = await page.$(selectors[workloadName]['id']);
+    await page.waitForSelector(selectors[workloadName]['id'] + idSuffix, {timeout: 10*1000});
+    let element = await page.$(selectors[workloadName]['id'] + idSuffix);
     console.log(`Downloading trends image for ${workloadName}`);
     await element.screenshot({
         path: path.join(process.cwd(), 'charts', `${i++}-${prefixDate}-${workloadName}-trends.png`)
@@ -89,11 +92,12 @@ async function getChartFiles() {
 */
 async function cleanUpChartFiles() {
   let chartFiles = await getChartFiles();
-
-  for (let file of chartFiles) {
-    let absChartFile = path.join(process.cwd(), 'charts', file);
-    console.log(`Remove chart file: ${absChartFile}`);
-    await fs.promises.unlink(absChartFile);
+  if ( chartFiles.length !== 0 ) {
+    for (let file of chartFiles) {
+      let absChartFile = path.join(process.cwd(), 'charts', file);
+      console.log(`Remove chart file: ${absChartFile}`);
+      await fs.promises.unlink(absChartFile);
+    }
   }
 
   return Promise.resolve();
